digit_recognition_app_dec2025/requirements.txt
streamlit==1.42.1
requests>=2.28.1
streamlit-sal==0.2.0
Pillow>=10.0.0

# OpenAI is optional - only needed for AI chat responses
# Uncomment the line below if you want to use OpenAI API:
# openai>=1.60.0

# Digit Recognition Dependencies (Optional)
# Core ML packages
tensorflow>=2.10.0; platform_system != "Darwin" or platform_machine != "arm64"
tensorflow-macos>=2.10.0; platform_system == "Darwin" and platform_machine == "arm64"

# OCR packages (Optional - install what you need
easyocr>=1.6.0  # Recommended: Works well, easy to install
pytesseract>=0.3.10  # Optional: Requires system Tesseract installation
# paddleocr>=2.6.0  # Optional: Can be heavy, uncomment if needed
# paddlepaddle>=2.0.0  # Required if using PaddleOCR

# Object detection for multi-digit recognition
ultralytics>=8.0.0  # YOLO v8 for digit detection

# Image processing
opencv-python>=4.5.0

# Note: Tesseract OCR engine must be installed separately:
#   - Windows: https://github.com/UB-Mannheim/tesseract/wiki
#   - Mac: brew install tesseract
#   - Linux: sudo apt-get install tesseract-ocr

